v2.1.0: feature request: send PTC commands on the command line
This is a feature request from https://github.com/la5nta/pat/issues/187
Idea is to send PTC commands on the command line. This is also useful
for bookmarks in the Web GUI.

Quite a number of changes were needed, as the commands need to be sent
to the PTC controller before going into WA8DED hostmode, so in the
init phase. That makes it also necessary to change some lines in
Martin's Pat code. From the logic of Pat, the URL is passed to the
driver in the Dial phase. However, this is too late as the PTC is
already in host mode at that stage.

Changes:

- in modem.go: OpenModem has now an additional string parameter
cmdlineinit
- in modem.go: the commands are added to the build-in set of commands
sent to the PTC
- everywhere: white space changes made by LiteIDE.